demo/requirements.txt
pygame 
rpi-lgpio
